requirements.txt
streamlit==1.37.1
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
numpy==2.0.1
Pillow==10.4.0
transformers==4.41.2